+++ init.lua
@@ -1,7 +1,7 @@
 require("balm/type_casts")
 
 return {
-  version = "2026.6.2",
+  version = "2026.6.4",
   m = require("balm/modules"),
   Object = require("balm/object"),
   p = require("balm/persistence"),