+++ tests/utils_test.lua
@@ -1,4 +1,6 @@
 require("balm/tests/u/id_generator_test")
 require("balm/tests/u/id128_generator_test")
 require("balm/tests/u/string_buffer_test")
+require("balm/tests/u/timeline_test")
+require("balm/tests/u/tweener_test")
 require("balm/tests/u/u128_test")